Cloned mec@mec:FileOffsetBits
Cleanup: cast st_size when printing.

+++ util/neo_files.c
@@ -81,12 +81,12 @@ NEOERR *ne_load_file_len (const char *path, char **str, int *out_len)
   }
 
   if (s.st_size >= INT_MAX)
-      return nerr_raise (NERR_ASSERT, "File %s too large (%ld >= INT_MAX)",
-                         path, s.st_size);
+      return nerr_raise (NERR_ASSERT, "File %s too large (%lld >= INT_MAX)",
+                         path, (long long) s.st_size);
 
   if (s.st_size < 0)
-      return nerr_raise (NERR_ASSERT, "File %s size error? (%ld < 0)", path,
-                         s.st_size);
+      return nerr_raise (NERR_ASSERT, "File %s size error? (%lld < 0)", path,
+                         (long long) s.st_size);
 
   fd = open (path, O_RDONLY);
   if (fd == -1)
